+++ CHANGELOG.md
@@ -87,6 +87,7 @@
 * sdk: automatically resend event after NIP-42 authentication ([Yuki Kishimoto])
 * sdk: add `Connection::embedded_tor_with_path` ([Yuki Kishimoto])
 * connect: add `NostrConnect::status` ([Yuki Kishimoto])
+* connect: add pubkey in `NostrConnectSignerActions::approve` ([Yuki Kishimoto])
 * relay-builder: add NIP42 support ([Yuki Kishimoto])
 * relay-builder: add negentropy support ([Yuki Kishimoto])
 * relay-builder: add read/write policy plugins ([v0l])